+++ app/templates/nav-main.php
@@ -29,6 +29,9 @@
                 <a href="/parks/lumpy-ridge-trailhead.php">Lumpy Ridge Trailhead</a>
               </li>
               <li>
+                <a href="/parks/south-shore.php">South Shore Parkway</a>
+              </li>
+              <li>
                 <a href="/parks">All Parks</a>
               </li>
             </ul>